requirements.txt
-e .
aws_cdk.aws_ec2
aws_cdk.aws_elasticloadbalancingv2
aws_cdk.aws_route53
aws_cdk.aws_route53_targets
aws_cdk.aws_rds
aws_cdk.aws_secretsmanager
aws_cdk.aws_certificatemanager
aws_cdk.aws_ecs
aws_cdk.aws_iam
aws_cdk.aws_ecr
aws_cdk.aws_ecr_assets
